# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "wordondialogue@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/tools": {
            "get": {
                "tags": [
                    "Tools"
                ],
                "summary": "Get List Tools",
                "operationId": "2c6afc0de4988cabb3d95fd469a50c23",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tools"
                ],
                "summary": "Create tool",
                "operationId": "b634977798d2b3c8f08ca41ad0ab0fc5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "link": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "tags": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/tools/{tags}": {
            "get": {
                "tags": [
                    "Tools"
                ],
                "summary": "Get Detail Todo",
                "operationId": "359015fedb9802ea8fe11bd3210c1f21",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/tools/{id}": {
            "delete": {
                "tags": [
                    "Tools"
                ],
                "summary": "Delete tool",
                "operationId": "f9aba5edf179996cbb16caecfbabf2dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}